requirements.txt
# FastAPI & Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database & Auth
supabase>=2.0.0
psycopg-binary>=3.2.0

# AI & NLP
openai>=1.3.0
nltk>=3.8.1
textblob>=0.17.1

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1